Fix several issues in layer lock filtering (#335)
The test suite now checks that the example stacks have
valid lock file committed, and that some of the specific
details of those layer locks are as expected.

The mlx example has now also been updated with
a macOS-only layer to demonstrate omitting multiple
packages due to environment marker mismatches.

Closes #333